Move transact::protocol::* to protocol::*
This is an incremental step toward integrating transact fully into
libsawtooth. This removes the transact feature guard on these protocol
structs, with the exception of receipt; receipt needs further
refactoring so that it does not reference non-protocol code.
