+++ libsawtooth/src/protocol/batch.rs
