+++ libsawtooth/src/protocol/receipt.rs
