+++ libsawtooth/src/protocol/transaction.rs
